requirements.txt
quart
openai
python-dotenv
uvicorn
